logs_sender/pyproject.toml
[project]
name = "logs-sender"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.4.1",
    "psycopg[binary,pool]>=3.3.2",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.46",
]
